Item9427: Only update button state if cursor idle more than 500ms
Improves cursor responsiveness on large documents (>250KiB), especially
on slow PCs and IE users.

Tested on IE6, IE7, IE8, Opera 10.61, Chrome 6.0.490.1, Safari 5.0.1, FF3.6.7

PLEASE HELP TEST and provide feedback. Maybe the 500ms lag between
cursor position and button state updates is too unsettling.

If we decide that this is the case, I will make this latency a pref
var and we can default it to 100ms or so, and then users/sites with large
docs can increase the number.

git-svn-id: http://svn.foswiki.org/trunk/TinyMCEPlugin@8610 0b4bb1d4-4e5a-0410-9cc4-b2b747904278